+++ documentation/ChangeLog.md
@@ -11,6 +11,7 @@
   * However the documentation popup information will no longer be show for any content assist entries.
  * Removed `gocode` bundle in Goclipse as it was very out of date (closes #88).
    * Added button/tool in Tools pref page to automatically download gocode from Github. 
+ * Added button/tool in Tools pref page to automatically download Go oracle from Github. 
  * Fixed #105: Project Explorer does not expand GOROOT for Go 1.4
  * Editor Ctrl-click (hyperlink) now only uses Go Oracle for the operation, the previous hyperlink has now been fully removed due to being buggy.
   * Improved: for Go Oracle integration, if the tool terminates with a non-zero exit code, instead of showing an error dialog,